% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastore_operations.R
\name{delete_container}
\alias{delete_container}
\title{Deletes the specified container}
\usage{
delete_container(ContainerName)
}
\arguments{
\item{ContainerName}{The name of the container to delete.}
}
\description{
Deletes the specified container. Before you make a \code{DeleteContainer} request, delete any objects in the container or in any folders in the container. You can delete only empty containers.
}
\examples{

}
