% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastore_operations.R
\name{delete_container_policy}
\alias{delete_container_policy}
\title{Deletes the access policy that is associated with the specified container}
\usage{
delete_container_policy(
  ContainerName = "string"
)
}
\arguments{
\item{ContainerName}{[required] The name of the container that holds the policy.}
}
\description{
Deletes the access policy that is associated with the specified container.
}
\examples{

}
