% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastore_operations.R
\name{delete_cors_policy}
\alias{delete_cors_policy}
\title{Deletes the cross-origin resource sharing (CORS) configuration information that is set for the container}
\usage{
delete_cors_policy(
  ContainerName = "string"
)
}
\arguments{
\item{ContainerName}{[required] The name of the container to remove the policy from.}
}
\description{
Deletes the cross-origin resource sharing (CORS) configuration information that is set for the container.
}
\details{
To use this operation, you must have permission to perform the \code{MediaStore:DeleteCorsPolicy} action. The container owner has this permission by default and can grant this permission to others.
}
\examples{

}
