% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastore_operations.R
\name{delete_lifecycle_policy}
\alias{delete_lifecycle_policy}
\title{Removes an object lifecycle policy from a container}
\usage{
delete_lifecycle_policy(ContainerName)
}
\arguments{
\item{ContainerName}{[required] The name of the container that holds the object lifecycle policy.}
}
\description{
Removes an object lifecycle policy from a container.
}
\section{Accepted Parameters}{
\preformatted{delete_lifecycle_policy(
  ContainerName = "string"
)
}
}

