% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastore_operations.R
\name{describe_container}
\alias{describe_container}
\title{Retrieves the properties of the requested container}
\usage{
describe_container(
  ContainerName = "string"
)
}
\arguments{
\item{ContainerName}{The name of the container to query.}
}
\description{
Retrieves the properties of the requested container. This request is commonly used to retrieve the endpoint of a container. An endpoint is a value assigned by the service when a new container is created. A container's endpoint does not change after it has been assigned. The \code{DescribeContainer} request returns a single \code{Container} object based on \code{ContainerName}. To return all \code{Container} objects that are associated with a specified AWS account, use ListContainers.
}
\examples{

}
