% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastore_operations.R
\name{get_cors_policy}
\alias{get_cors_policy}
\title{Returns the cross-origin resource sharing (CORS) configuration information that is set for the container}
\usage{
get_cors_policy(ContainerName)
}
\arguments{
\item{ContainerName}{The name of the container that the policy is assigned to.}
}
\description{
Returns the cross-origin resource sharing (CORS) configuration information that is set for the container.
}
\details{
To use this operation, you must have permission to perform the \code{MediaStore:GetCorsPolicy} action. By default, the container owner has this permission and can grant it to others.
}
\examples{

}
