% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastore_operations.R
\name{get_lifecycle_policy}
\alias{get_lifecycle_policy}
\title{Retrieves the object lifecycle policy that is assigned to a container}
\usage{
get_lifecycle_policy(
  ContainerName = "string"
)
}
\arguments{
\item{ContainerName}{[required] The name of the container that the object lifecycle policy is assigned to.}
}
\description{
Retrieves the object lifecycle policy that is assigned to a container.
}
\examples{

}
