% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastore_operations.R
\name{list_containers}
\alias{list_containers}
\title{Lists the properties of all containers in AWS Elemental MediaStore}
\usage{
list_containers(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{Only if you used \code{MaxResults} in the first command, enter the token (which was included in the previous response) to obtain the next set of containers. This token is included in a response only if there actually are more containers to list.}

\item{MaxResults}{Enter the maximum number of containers in the response. Use from 1 to 255 characters.}
}
\description{
Lists the properties of all containers in AWS Elemental MediaStore.
}
\details{
You can query to receive all the containers in one response. Or you can include the \code{MaxResults} parameter to receive a limited number of containers in each response. In this case, the response includes a token. To get the next set of containers, send the command again, this time with the \code{NextToken} parameter (with the returned token as its value). The next set of responses appears, with a token if there are still more containers to receive.

See also DescribeContainer, which gets the properties of one container.
}
\examples{

}
