% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastore_package.R
\docType{package}
\name{paws.mediastore-package}
\alias{paws.mediastore}
\alias{paws.mediastore-package}
\title{paws.mediastore: AWS Elemental MediaStore}
\description{
An AWS Elemental MediaStore container is a namespace that
holds folders and objects. You use a container endpoint to create,
read, and delete objects.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
