% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastore_operations.R
\name{put_container_policy}
\alias{put_container_policy}
\title{Creates an access policy for the specified container to restrict the users and clients that can access it}
\usage{
put_container_policy(ContainerName, Policy)
}
\arguments{
\item{ContainerName}{The name of the container.}

\item{Policy}{The contents of the policy, which includes the following:
\itemize{
\item One \code{Version} tag
\item One \code{Statement} tag that contains the standard tags for the policy.
}}
}
\description{
Creates an access policy for the specified container to restrict the users and clients that can access it. For information about the data that is included in an access policy, see the \href{https://aws.amazon.com/documentation/iam/}{AWS Identity and Access Management User Guide}.
}
\details{
For this release of the REST API, you can create only one policy for a container. If you enter \code{PutContainerPolicy} twice, the second command modifies the existing policy.
}
\examples{

}
