% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastore_operations.R
\name{put_lifecycle_policy}
\alias{put_lifecycle_policy}
\title{Writes an object lifecycle policy to a container}
\usage{
put_lifecycle_policy(ContainerName, LifecyclePolicy)
}
\arguments{
\item{ContainerName}{[required] The name of the container that you want to assign the object lifecycle policy to.}

\item{LifecyclePolicy}{[required] The object lifecycle policy to apply to the container.}
}
\description{
Writes an object lifecycle policy to a container. If the container already has an object lifecycle policy, the service replaces the existing policy with the new policy.
}
\section{Accepted Parameters}{
\preformatted{put_lifecycle_policy(
  ContainerName = "string",
  LifecyclePolicy = "string"
)
}
}

